updating to latest version of molfile plugin

+++ vmd/plugins/molfile_plugin/src/endianswap.h
@@ -10,7 +10,7 @@
  *
  *      $RCSfile: endianswap.h,v $
  *      $Author: johns $       $Locker:  $             $State: Exp $
- *      $Revision: 1.7 $       $Date: 2016/11/28 05:01:53 $
+ *      $Revision: 1.8 $       $Date: 2020/10/21 18:03:15 $
  *
  ***************************************************************************
  * DESCRIPTION:
@@ -28,9 +28,11 @@
 #ifndef ENDIAN_SWAP_H
 #define ENDIAN_SWAP_H
 
+#include <stddef.h>
+
 /* works on unaligned 2-byte quantities */
-static void swap2_unaligned(void *v, long ndata) {
-  long i;
+static void swap2_unaligned(void *v, ptrdiff_t ndata) {
+  ptrdiff_t i;
   char * dataptr = (char *) v;
   char tmp;
 
@@ -43,8 +45,8 @@ static void swap2_unaligned(void *v, long ndata) {
 
 
 /* works on unaligned 4-byte quantities */
-static void swap4_unaligned(void *v, long ndata) {
-  long i;
+static void swap4_unaligned(void *v, ptrdiff_t ndata) {
+  ptrdiff_t i;
   char *dataptr;
   char tmp;
 
@@ -62,9 +64,9 @@ static void swap4_unaligned(void *v, long ndata) {
 
 
 /* works on unaligned 8-byte quantities */
-static void swap8_unaligned(void *v, long ndata) {
+static void swap8_unaligned(void *v, ptrdiff_t ndata) {
   char *data = (char *) v;
-  long i;
+  ptrdiff_t i;
   char byteArray[8];
   char *bytePointer;
 
@@ -93,9 +95,9 @@ static void swap8_unaligned(void *v, long ndata) {
 
 /* Only works with aligned 2-byte quantities, will cause a bus error */
 /* on some platforms if used on unaligned data.                      */
-static void swap2_aligned(void *v, long ndata) {
+static void swap2_aligned(void *v, ptrdiff_t ndata) {
   short *data = (short *) v;
-  long i;
+  ptrdiff_t i;
   short *N; 
 
   for (i=0; i<ndata; i++) {
@@ -107,9 +109,9 @@ static void swap2_aligned(void *v, long ndata) {
 
 /* Only works with aligned 4-byte quantities, will cause a bus error */
 /* on some platforms if used on unaligned data.                      */
-static void swap4_aligned(void *v, long ndata) {
+static void swap4_aligned(void *v, ptrdiff_t ndata) {
   int *data = (int *) v;
-  long i;
+  ptrdiff_t i;
   int *N;
   for (i=0; i<ndata; i++) {
     N = data + i;
@@ -121,12 +123,12 @@ static void swap4_aligned(void *v, long ndata) {
 
 /* Only works with aligned 8-byte quantities, will cause a bus error */
 /* on some platforms if used on unaligned data.                      */
-static void swap8_aligned(void *v, long ndata) {
+static void swap8_aligned(void *v, ptrdiff_t ndata) {
   /* Use int* internally to prevent bugs caused by some compilers */
   /* and hardware that would potentially load data into an FP reg */
   /* and hose everything, such as the old "jmemcpy()" bug in NAMD */
   int *data = (int *) v;  
-  long i;
+  ptrdiff_t i;
   int *N; 
   int t0, t1;
